make working package
